Add Priors to database_schema module
