+++ ert_storage/database_schema/__init__.py
@@ -1,0 +1,8 @@
+from .metadatafield import MetadataField
+from .record import Record, F64Matrix, File, FileBlock, RecordClass, RecordType
+from .ensemble import Ensemble
+from .experiment import Experiment
+from .observation import Observation, ObservationTransformation
+from .update import Update
+from .prior import Prior, PriorFunction
+from ert_storage.database import Base
