+++ setup.py
@@ -13,6 +13,7 @@ setup(
         "ert_storage._alembic",
         "ert_storage._alembic.alembic",
         "ert_storage._alembic.alembic.versions",
+        "ert_storage.database_schema",
         "ert_storage.endpoints",
         "ert_storage.endpoints.compute",
         "ert_storage.compute",
